% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{dc_loglik}
\alias{dc_loglik}
\title{Compute the log likelihood for a set of dual censored observations}
\usage{
dc_loglik(
  data,
  z = "ATRTN",
  fns = list(int12 = function(x, z) 0, int02 = function(x, z) 0, P00 = function(l, r, z)
    0, logP00 = function(l, r, z) 1, P01 = function(l, r, z) 0, P11 = function(l, r, z)
    0, logP11 = function(l, r, z) 1)
)
}
\arguments{
\item{data}{data frame containing required columns, see details}

\item{fns}{a list of functions to compute the likelihood}
}
\value{
vector of log likelihoods
}
\description{
Compute the log likelihood for a set of dual censored observations
}
\details{
`data` must contain columns called `l` (optionally `left`), `r` (optionally `right`),
`v`, and either `dX`, `d0X`,`deltaX` or `delta0X` for all `x` \in (1,2,3).

`fns` must contain entries `int12`, `int02`, `P00`, `logP00`, `P01`, `P11`,
`logP11`. These functions should be able to take a vector of inputs.
}
