% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsm_mod.R
\name{dcsm_se}
\alias{dcsm_se}
\title{Return a function that calculate the standard error of the survivor curve}
\usage{
dcsm_se(x, fns, tol = NULL)
}
\arguments{
\item{x}{A \code{dscm_mod} object}

\item{fns}{List of functions used to generate SE}
}
\value{
A function to find the standard error at time \code{tm} with treatment arm \code{z}.
}
\description{
Return a function that calculate the standard error of the survivor curve
}
