% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joly_penalised.R
\name{joly.fit}
\alias{joly.fit}
\title{Fit a dataset to a model with I-spline baseline transition intensities.}
\usage{
joly.fit(
  data,
  k01,
  k02,
  k12,
  kappa01 = 1,
  kappa02 = 1,
  kappa12 = 1,
  compute_cross = T,
  debug = F,
  initials = NULL,
  control = list(fnscale = -1, maxit = 500),
  method = "BFGS",
  ...
)
}
\arguments{
\item{data}{Data set to model from}

\item{k01}{Number of interior knots in the 0 -> 1 transition}

\item{k02}{Number of interior knots in the 0 -> 2 transition}

\item{k12}{Number of interior knots in the 1 -> 2 transition}

\item{kappa01}{Penalty parameter for the 0 -> 1 transition}

\item{kappa02}{Penalty parameter for the 0 -> 1 transition}

\item{kappa12}{Penalty parameter for the 0 -> 1 transition}

\item{compute_cross}{Compute the cross value (and relevant hessians)? EXPENSIVE!}

\item{debug}{Print every imputation?}

\item{initials}{optional list of initial values to replace call to \link{joly.initials}}
}
\value{
A \code{optim} object.
}
\description{
Requires known interior knots and penalty parameters.
}
