% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joly_penalised.R
\name{joly.initials}
\alias{joly.initials}
\title{Get a set of feasible initial values for a I-spline based model}
\usage{
joly.initials(data, k01, k02, k12)
}
\arguments{
\item{data}{Data set to model from, with columns `L`, `R`, `V`, `Delta1`, `Delta2`}

\item{k01}{Number of interior knots in the 0 -> 1 transition}

\item{k02}{Number of interior knots in the 0 -> 2 transition}

\item{k12}{Number of interior knots in the 1 -> 2 transition}
}
\value{
A list of parameters, to be fed to \link{joly.ll}
}
\description{
Get a set of feasible initial values for a I-spline based model
}
