% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joly_penalised.R
\name{joly.ll}
\alias{joly.ll}
\title{Compute the log likelihood for a model with illness death and I-spline cumulative intensities}
\usage{
joly.ll(p, data, indiv = FALSE)
}
\arguments{
\item{p}{parameters, as a named list based off \link{Joly}}

\item{data}{dataset as per instructions in \link{dc_loglik}}

\item{indiv}{return a vector of each individual log-likelihood instead of the sum?}
}
\value{
The log-likelihood for the data and parameters provided.
}
\description{
Compute the log likelihood for a model with illness death and I-spline cumulative intensities
}
