% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joly_penalised.R
\name{joly.fnBuilder}
\alias{joly.fnBuilder}
\title{Functions for a illness-death model with I-spline cumulative intensities}
\usage{
joly.fnBuilder(
  theta01,
  theta02,
  theta12,
  gammas01,
  knots01,
  gammas02,
  knots02,
  gammas12,
  knots12,
  boundaries
)
}
\arguments{
\item{theta01}{A single numeric value representing treatment difference in the progression intensity}

\item{theta02}{A single numeric value representing treatment difference in the pre-progression death intensity}

\item{theta12}{A single numeric value representing treatment difference in the post-progression death- intensity}

\item{gammas01}{Gamma values for the progression intensity spline}

\item{knots01}{Interior knots for the progression intensity spline}

\item{gammas02}{Gamma values for the pre-progression death intensity spline}

\item{knots02}{Interior knots for the pre-progression death intensity spline}

\item{gammas12}{Gamma values for the post-progression death intensity spline}

\item{knots12}{Interior knots for the post-progression death intensity spline}

\item{boundaries}{Boundary knots for the data (typically day 1 and the administrative censoring time). Cannot be 0.}
}
\value{
A list of functions.
}
\description{
Based on research by Joly et al. (1998, 2002)
}
\details{
Each transition's cumulative intensity is modeled as
\deqn{
 \text{A}_{hj} = S_{hj}(t;\boldsymbol{\gamma})\exp(z\theta_{hj})
}
\code{gammasXX} must contain the \eqn{\boldsymbol{\gamma}_{hj}} values as parameters.
Using an example, if we have one knotpoint in the transition from 0 to 1: \code{gamma010}, \code{gamma011}, \code{gamma012} must be provided as parameters.
}
