% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pars.R
\name{make_pars}
\alias{make_pars}
\title{Bundle a vector of parameters into a list of parameters}
\usage{
make_pars(p)
}
\arguments{
\item{p}{A named vector of parameters, typically from a \link{joly.fit} or
\link{royston_parmar.fit} call.}
}
\value{
A list with entries \code{theta01}, \code{theta02}, \code{theta12},
\code{gammas01}, \code{knots01} \code{gammas02}, \code{knots02},
\code{gammas12}, \code{knots12}, \code{boundaries}.
}
\description{
Bundle a vector of parameters into a list of parameters
}
