% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pars.R
\name{make_pars2}
\alias{make_pars2}
\title{Update an initial set of parameters with new value}
\usage{
make_pars2(p, initials)
}
\arguments{
\item{p}{Named numeric vector of new values}

\item{inits}{List of initial values, typically from \link{royston_parmar.initials} or \link{joly.initials}}
}
\value{
a list
}
\description{
Update an initial set of parameters with new value
}
