% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nonlinear.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Perform a simple Newton-Raphson procedure for generic f and df functions.}
\usage{
newton_raphson(
  f = function(x, z) (x - 1)^2,
  df = function(x, z) 2 * (x - 1),
  z,
  t1 = 1
)
}
\arguments{
\item{f}{function taking two inputs, \code{x} and \code{z}.}

\item{df}{first derivative of function \code{f}.}

\item{z}{treatment arm}

\item{t1}{Starting value for \code{f} and \code{df}}
}
\value{
single numeric of the true \eqn{t}
}
\description{
Perform a simple Newton-Raphson procedure for generic f and df functions.
}
