% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{P01}
\alias{P01}
\title{Generic-ish function to compute the 0 \eqn{\to} 1 transition probability}
\usage{
P01(
  l,
  r,
  z,
  type,
  theta01_,
  theta02_,
  theta12_,
  gammas01,
  knots01,
  gammas02,
  knots02,
  gammas12,
  knots12,
  boundaries
)
}
\arguments{
\item{l}{The lower bounds of the (untransformed) interval}

\item{r}{The upper bounds of the (untransformed) interval}

\item{z}{The arms (case/control etc...)}

\item{type}{A string of either \code{"RoystonParmar"} or \code{"JolyPenalised"}}

\item{theta01_, theta02_, theta12_, gammas01, knots01, gammas02, knots02, gammas12, knots12, boundaries}{Passed to the \link{IllnessDeath} constructors to create the relevant set of functions.}
}
\description{
Typically wrapped by a separate function so that only the first 3 arguments are provided.
}
\details{
If \code{type = "RoystonParmar"} is used then a transformed interval with the substitution \eqn{u = exp(v)} is used to avoid a discontinuity at 0.
This then becomes an infinite integral.
}
