% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsm_mod.R
\name{plot.dcsm_mod}
\alias{plot.dcsm_mod}
\title{Plot the survivor curve for a dual-censored model}
\usage{
\method{plot}{dcsm_mod}(x, treat_col = "blue", add = F, ...)
}
\arguments{
\item{x}{A \code{dcsm_mod} object}

\item{treat_col}{The colour of the treatment effect line, set to \code{NULL} to only plot the baseline survivor function.}

\item{add}{Add this to a existing plot?}

\item{...}{Passed to \link{curve}}
}
\value{
Nothing, plots to output
}
\description{
Plot the survivor curve for a dual-censored model
}
