% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsm_mod.R
\name{print.dcsm_mod}
\alias{print.dcsm_mod}
\title{Print the treatment effect estimates for a dual-censored model}
\usage{
\method{print}{dcsm_mod}(x, ...)
}
\arguments{
\item{x}{A \code{dcsm_mod} object}

\item{...}{Not used}
}
\value{
Nothing, prints three estimates of the treatment effect for each transition
}
\description{
Print the treatment effect estimates for a dual-censored model
}
