% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royston_parmar.R
\name{royston_parmar.fit}
\alias{royston_parmar.fit}
\title{Fit a dataset to Royston & Parmar spline model.}
\usage{
royston_parmar.fit(
  data,
  k01,
  k02,
  k12,
  debug = F,
  initials = NULL,
  control = list(fnscale = -1, maxit = 500),
  method = "BFGS",
  ...
)
}
\arguments{
\item{data}{Data set to model from}

\item{k01}{Number of interior knots in the 0 -> 1 transition}

\item{k02}{Number of interior knots in the 0 -> 2 transition}

\item{k12}{Number of interior knots in the 1 -> 2 transition}

\item{debug}{Print every parameter tried?}

\item{initials}{Alternative list of initial values to try}

\item{control}{Passed to \link{optim}}

\item{method}{Passed to \link{optim}, defaults to \code{"BFGS"}}

\item{...}{Passed to \link{optim}}
}
\value{
A \link{optim} object.
}
\description{
Fit a dataset to Royston & Parmar spline model.
}
