% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royston_parmar.R
\name{royston_parmar.ll}
\alias{royston_parmar.ll}
\title{Compute the log likelihood for a model with illness death and natural cubic spline intensities}
\usage{
royston_parmar.ll(p, data, indiv = FALSE)
}
\arguments{
\item{p}{parameters, as a named list based off \link{RoystonParmar}}

\item{data}{dataset as per instructions in \link{dc_loglik}}

\item{indiv}{return a vector of each individual log-likelihood instead of the sum?}
}
\value{
The log-likelihood for the data and parameters provided.
}
\description{
Compute the log likelihood for a model with illness death and natural cubic spline intensities
}
