% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royston_parmar.R
\name{royston_parmar.fnBuilder}
\alias{royston_parmar.fnBuilder}
\title{Functions for a illness-death model with natural cubic spline intensities}
\usage{
royston_parmar.fnBuilder(
  theta01,
  theta02,
  theta12,
  gammas01,
  knots01,
  gammas02,
  knots02,
  gammas12,
  knots12,
  boundaries
)
}
\arguments{
\item{theta01}{A single numeric value representing treatment difference in the progression intensity}

\item{theta02}{A single numeric value representing treatment difference in the pre-progression death intensity}

\item{theta12}{A single numeric value representing treatment difference in the post-progression death- intensity}

\item{gammas01}{Gamma values for the progression intensity spline}

\item{knots01}{Interior knots for the progression intensity spline}

\item{gammas02}{Gamma values for the pre-progression death intensity spline}

\item{knots02}{Interior knots for the pre-progression death intensity spline}

\item{gammas12}{Gamma values for the post-progression death intensity spline}

\item{knots12}{Interior knots for the post-progression death intensity spline}

\item{boundaries}{Boundary knots for the data (typically day 1 and the administrative censoring time). Cannot be 0.}
}
\value{
A list of functions.
}
\description{
Based on research by Royston & Parmar (2002)
}
\details{
Each transition's baseline intensity is modeled as \eqn{\exp(S_{hj}(x;\boldsymbol{\gamma}_{hj}))}, where \eqn{x = \log(t)}.
We define \eqn{S_{hj}(x)} as a natural cubic spline such that it can be represented in terms of its basis functions \eqn{v_j(x)}
\deqn{
 S_{hj}(x) = \gamma_{hj0} + \gamma_{hj1}x + \sum^K_{i=1} \gamma_{hji} v_j(x)
}
with \eqn{K} the number of internal knot points.
\code{gammasXX} must contain the \eqn{\boldsymbol{\gamma}_{hj}} values as parameters.
Using an example, if we have one knotpoint in the transition from 0 to 1: \code{gamma010}, \code{gamma011}, \code{gamma012} must be provided as parameters.
}
