% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exp.R
\name{sim_exp}
\alias{sim_exp}
\title{Simulate data based on exponential times to event}
\usage{
sim_exp(
  N = 300,
  lambda01 = 4,
  lambda02 = 1,
  lambda12 = 1.5,
  theta01 = log(0.4),
  theta02 = log(0.4),
  theta12 = log(1),
  pct_no_prg = 0.05,
  K = 8
)
}
\arguments{
\item{N}{The number of observations to simulate}

\item{lambda01}{The rate for time-to-progression}

\item{lambda02}{The rate for time-to-death pre-progression}

\item{lambda12}{The rate for time-to-death post-progression}

\item{theta01}{The treatment effect for time-to-progression}

\item{theta02}{The treatment effect for time-to-death pre-progression}

\item{theta12}{The treatment effect for time-to-death post-progression}

\item{pct_no_prg}{The percent of observations assumed to have not progressed}

\item{K}{The number of visits}
}
\value{
A data frame with 13 columns: \itemize{
 \item `T01`, `T02`, `T12` - The 'true' times of the event based on the `lambdaXX` parameter.
 \item `PFS` - Did the subject progress or die before censoring?
 \item `PFSDY` - The day of progression based on when it was first observed.
 \item `TRUEDY` - The true day of progression.
 \item `L`, `R` - Either the interval progression occurs in if `delta1 == 1`, or the last interval prior to censoring.
 \item `V` - Either the time of death if `delta2 == 1`, or the time of right-censoring.
 \item `delta0` - Did the patient definitely not progress?
 \item `delta1` - Did the patient definitely progress?
 \item `delta2` - Did the patient die?
 \item `ATRTN` - `0` => control arm, `1` => active arm.
}
}
\description{
Simulate data based on exponential times to event
}
\details{
\eqn{T_{01} \sim \text{Exp}(\lambda_{01})} and
\eqn{T_{02} \sim \text{Exp}(\lambda_{02})} are drawn first. If \eqn{T_{01}}
is less than \eqn{T_02} then
\eqn{T_12 \sim T_01 + \text{Exp}(\lambda_{12} \equiv \text{Gamma}(2, \lambda_{01} \times \lambda_{02}))}.
These true values are then used to find \eqn{C_a}, the administrative
censoring time as the 80th percentile of \eqn{D = T_{02} \vee T_{12}}.
}
