% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royston_parmar.R
\name{solve_qp_problem}
\alias{solve_qp_problem}
\title{Find initial coefficients for Royston-Parmar
Taken nearly directly from flexsurv (c. Christopher Jackson)}
\usage{
solve_qp_problem(y, x, z, knots, boundaries)
}
\arguments{
\item{y}{cumhaz}

\item{x}{log vector of time (R/V)}

\item{z}{vector of coefficients}

\item{initial}{initial values so far (for knots/boundaries)}
}
\value{
vector of solutions
}
\description{
Find initial coefficients for Royston-Parmar
Taken nearly directly from flexsurv (c. Christopher Jackson)
}
