% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weib.fit}
\alias{weib.fit}
\title{Fit a model based on starting parameters}
\usage{
weib.fit(
  data,
  initial = c(lambda01 = 1, gamma01 = 1, theta01 = 0, lambda02 = 1, gamma02 = 1, theta02
    = 0, lambda12 = 1, gamma12 = 1, theta12 = 0),
  control = list(fnscale = -1, maxit = 500),
  method = "BFGS",
  ...
)
}
\arguments{
\item{data}{Data set to fit to, must have entries as described in \link{dc_loglik}}

\item{initial}{Vector of initial parameters}

\item{control}{passed to \link{optim}, defaults to change the scale of the log-likelihood function}

\item{method}{passed to \link{optim}, defaults to \code{"Nelder-Mead"}.}

\item{...}{additional parameters passed to \link{optim}}
}
\value{
A \link{optim} output
}
\description{
Fit a model based on starting parameters
}
