% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weib.ll}
\alias{weib.ll}
\title{Compute the log likelihood for a model with illness death and Weibull intensities}
\usage{
weib.ll(p, data)
}
\arguments{
\item{p}{parameters, see \link{weibull}}

\item{data}{dataset as per instructions in \link{dc_loglik}}
}
\value{
The log-likelihood for the data and parameters provided.
}
\description{
Compute the log likelihood for a model with illness death and Weibull intensities
}
