% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weib.fnBuilder}
\alias{weib.fnBuilder}
\alias{weib.int}
\alias{weib.cumint}
\alias{weib.P00}
\alias{weib.logP00}
\alias{weib.P11}
\alias{weib.logP11}
\alias{weib.P01}
\title{Functions for a illness-death model with Weibull intensities}
\usage{
weib.fnBuilder(
  lambda01,
  gamma01,
  theta01,
  lambda02,
  gamma02,
  theta02,
  lambda12,
  gamma12,
  theta12
)

weib.int(lambda, gamma, theta)

weib.cumint(lambda, gamma, theta)

weib.P00(lambda01, gamma01, theta01, lambda02, gamma02, theta02)

weib.logP00(lambda01, gamma01, theta01, lambda02, gamma02, theta02)

weib.P11(lambda12, gamma12, theta12)

weib.logP11(lambda12, gamma12, theta12)

weib.P01(
  lambda01,
  gamma01,
  theta01,
  lambda02,
  gamma02,
  theta02,
  lambda12,
  gamma12,
  theta12
)
}
\arguments{
\item{lambda}{see details, can be suffixed based on the transition}

\item{gamma}{see details, can be suffixed based on the transition}

\item{theta}{see details, can be suffixed based on the transition}
}
\value{
A list of functions, typically used in `?weib.ll`
}
\description{
Functions for a illness-death model with Weibull intensities
}
\details{
Weibull is formulated such that its hazard is
\deqn{
H(t) = \lambda \gamma t^{\gamma-1}
}
this then models the baseline intensity function for each transition \eqn{h \to j}
\deqn{
\alpha_{hj}(t;z) = \alpha_{hj}(t;0) \exp(z\theta_{hj})
}
}
